@@ -6,7 +6,9 @@
     "dev": "next dev",
     "build": "next build",
     "start": "next start",
-    "lint": "next lint"
+    "lint": "eslint .",
+    "lint:fix": "eslint . --fix",
+    "format": "prettier --write ."
   },
   "dependencies": {
     "@apollo/client": "^3.8.9",
@@ -22,6 +24,7 @@
     "react-dom": "^18",
     "react-redux": "^9.0.4",
     "react-toast": "^1.0.3",
+    "react-toastify": "^9.1.3",
     "redux-persist": "^6.0.0",
     "uuid": "^9.0.1"
   },
@@ -31,9 +34,14 @@
     "@types/react": "18.2.47",
     "@types/uuid": "^9.0.7",
     "autoprefixer": "^10.0.1",
-    "eslint": "^8",
+    "eslint": "^8.56.0",
     "eslint-config-next": "14.0.4",
+    "eslint-config-prettier": "^9.1.0",
+    "eslint-plugin-prettier": "^5.1.3",
+    "eslint-plugin-react": "^7.33.2",
+    "eslint-plugin-react-hooks": "^4.6.0",
     "postcss": "^8",
+    "prettier": "^3.2.1",
     "tailwindcss": "^3.3.0"
   }
 }
