@@ -1,11 +1,7 @@
-"use client"
+"use client";
 import store from "./store/store";
 import { Provider } from "react-redux";
 
-export function ReduxProvider({children}) {
-    return (
-        <Provider store={store}>
-            {children}
-        </Provider>
-    );
+export function ReduxProvider({ children }) {
+  return <Provider store={store}>{children}</Provider>;
 }
